Updated assign button
It will now try to assign the item with the same runeslots of the one the player linked during bid phase. If that item has already been assigned (or anyway can't be assigned) it'll try to assign an equivalent item (in case of multiple drops of the same item).
This feature has not been tested yet as It's quite hard to.